✏️  minor ui fixes/ code refactoring datatable pages
